requirements.txt
torch
transformers
underthesea
vncorenlp
pandas
numpy
pillow
datasets